requirements.txt
line-bot-sdk
flask
openai < 1.0.0
gunicorn
Flask==2.1.3
psycopg2-binary==2.9.3
SQLAlchemy==1.4.27
tiktoken==0.1.2
langchain-openai==0.0.2
langchain-core==0.1.7
configparser==5.2.0

